+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `threeDSecureAuthenticationID` to `CreditCardVerificationRequest`
 * Add `threeDSecureToken` to `CreditCardVerificationRequest`
 * Add `verificationThreeDSecurePassThruRequest` to `CreditCardVerificationRequest`
+* Add subscriptions to `SepaDirectDebitAccount`
 
 ## 3.20.0
 * Add `SEPADirectDebitAccount` payment method